rename and export in index
