+++ src/index.ts
@@ -1,16 +1,33 @@
-export { default as useMount } from "./useMount";
-export { default as useUnmount } from "./useUnmount";
-export { default as useResource } from "./useResource";
-export { default as useRerender } from "./useRerender";
-
+import useCurriedState from './useCurriedState';
+import useForceUpdate from "./useForceUpdate";
+import useFreeze from './useFreeze';
+import useKeybind from './useKeybind';
 import useMount from "./useMount";
-import useUnmount from "./useUnmount";
+import useRefState from './useRefState';
 import useResource from "./useResource";
-import useRerender from "./useRerender";
+import useStealthState from './useStealthState';
+import useUnmount from "./useUnmount";
+
+export {
+  useCurriedState,
+  useForceUpdate,
+  useFreeze,
+  useKeybind,
+  useMount,
+  useRefState,
+  useResource,
+  useStealthState,
+  useUnmount,
+};
 
 export default {
+  useCurriedState,
+  useForceUpdate,
+  useFreeze,
+  useKeybind,
   useMount,
-  useUnmount,
+  useRefState,
   useResource,
-  useRerender,
+  useStealthState,
+  useUnmount,
 };
