+++ src/useStealthState.ts
@@ -1,0 +1,7 @@
+import { useRef, useState } from "react";
+
+export default ((init?) => {
+  const ref = useRef(init);
+
+  return [ref.current, (newVal) => (ref.current = newVal)];
+}) as typeof useState;
